fix: create ApiModels folder and separate in and out models
