+++ Doppler.BillingUser/ApiModels/PaymentStatusApiEnum.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel;
+
+namespace Doppler.BillingUser.ApiModels
+{
+    public enum PaymentStatusApiEnum
+    {
+        Approved,
+        Pending,
+        Declined
+    }
+}
