--- Doppler.BillingUser/Enums/ReprocessInvoicePaymentResultEnum.cs
@@ -1,8 +1,0 @@
-namespace Doppler.BillingUser.Enums
-{
-    public enum ReprocessInvoicePaymentResultEnum
-    {
-        Successful,
-        Failed
-    }
-}
