+++ Doppler.BillingUser/Mappers/PaymentStatus/IPaymentStatusMapper.cs
@@ -1,3 +1,4 @@
+using Doppler.BillingUser.ApiModels;
 using Doppler.BillingUser.Enums;
 
 namespace Doppler.BillingUser.Mappers.PaymentStatus
@@ -5,5 +6,7 @@ namespace Doppler.BillingUser.Mappers.PaymentStatus
     public interface IPaymentStatusMapper
     {
         PaymentStatusEnum MapToPaymentStatus(MercadoPagoPaymentStatusEnum status);
+
+        PaymentStatusEnum MapFromPaymentStatusApiEnumToPaymentStatusEnum(PaymentStatusApiEnum status);
     }
 }
