+++ Doppler.BillingUser/Mappers/PaymentStatus/PaymentStatusMapper.cs
@@ -1,9 +1,21 @@
+using Doppler.BillingUser.ApiModels;
 using Doppler.BillingUser.Enums;
+using System.Collections;
 
 namespace Doppler.BillingUser.Mappers.PaymentStatus
 {
     public class PaymentStatusMapper : IPaymentStatusMapper
     {
+        public PaymentStatusEnum MapFromPaymentStatusApiEnumToPaymentStatusEnum(PaymentStatusApiEnum status)
+        {
+            return status switch
+            {
+                PaymentStatusApiEnum.Approved => PaymentStatusEnum.Approved,
+                PaymentStatusApiEnum.Pending => PaymentStatusEnum.Pending,
+                PaymentStatusApiEnum.Declined => PaymentStatusEnum.DeclinedPaymentTransaction
+            };
+        }
+
         public PaymentStatusEnum MapToPaymentStatus(MercadoPagoPaymentStatusEnum status)
         {
             return status switch
